requirements.txt
fastapi==0.109.0
uvicorn==0.27.0 
pinecone-client==3.1.0 #VectorDB driver
sentence-transformers==2.3.1
google-generativeai>=0.7.0
pydantic==2.6.0
python-multipart
httpx
